refactored header and memoized. Minor telemetry refactoring and memoization in image segmentation
