+++ src/components/OHAEditor/index.js
@@ -30,6 +30,7 @@ import useTimeToCompleteSample from "../../utils/use-time-to-complete-sample.js"
 import TextField from "@material-ui/core/TextField"
 import { useToasts } from "../Toasts"
 import { setIn, without } from "seamless-immutable"
+import useEventCallback from "use-event-callback"
 
 import "brace/mode/javascript"
 import "brace/theme/github"
@@ -51,6 +52,8 @@ const useStyles = makeStyles({
     height: "100vh"
   }
 })
+
+const headerTabs = ["Settings", "Samples", "Label"]
 
 export default ({
   datasetName = "Universal Data Tool",
@@ -106,6 +109,8 @@ export default ({
     } catch (e) {}
   }, [jsonText])
 
+  const onChangeTab = useEventCallback(tab => changeMode(tab.toLowerCase()))
+
   let percentComplete = 0
   if (
     oha.taskOutput &&
@@ -136,9 +141,9 @@ export default ({
             />
           )
         }
-        onChangeTab={tab => changeMode(tab.toLowerCase())}
+        onChangeTab={onChangeTab}
         currentTab={mode}
-        tabs={["Settings", "Samples", "Label"]}
+        tabs={headerTabs}
       />
       <div style={{ overflowY: "scroll" }}>
         {mode === "json" && (
@@ -217,19 +222,31 @@ export default ({
             onSaveTaskOutputItem={(relativeIndex, output) => {
               let newOHA = oha
               if (!newOHA.taskOutput) {
-                newOHA = setIn(newOHA, ["taskOutput"], (newOHA.taskData || []).map(td => null))
+                newOHA = setIn(
+                  newOHA,
+                  ["taskOutput"],
+                  (newOHA.taskData || []).map(td => null)
+                )
               }
               if (
                 newOHA.taskOutput.length < newOHA.taskData.length ||
                 newOHA.taskOutput.includes(undefined)
               ) {
-                newOHA = setIn(newOHA, ["taskOutput"], newOHA.taskData.map((td, tdi) =>
-                  newOHA.taskOutput[tdi] === undefined
-                    ? null
-                    : newOHA.taskOutput[tdi]
-                ))
+                newOHA = setIn(
+                  newOHA,
+                  ["taskOutput"],
+                  newOHA.taskData.map((td, tdi) =>
+                    newOHA.taskOutput[tdi] === undefined
+                      ? null
+                      : newOHA.taskOutput[tdi]
+                  )
+                )
               }
-              newOHA = setIn(newOHA, ["taskOutput", singleSampleOHA.sampleIndex], output)
+              newOHA = setIn(
+                newOHA,
+                ["taskOutput", singleSampleOHA.sampleIndex],
+                output
+              )
               onChangeOHA(newOHA)
               changeSingleSampleOHA(null)
               if (singleSampleOHA.startTime) {
@@ -296,7 +313,9 @@ export default ({
           changeSampleInputEditor({ open: false })
         }}
         onChange={newInput => {
-          onChangeOHA(setIn(oha, ["taskData", sampleInputEditor.sampleIndex], newInput))
+          onChangeOHA(
+            setIn(oha, ["taskData", sampleInputEditor.sampleIndex], newInput)
+          )
         }}
       />
     </div>
