+++ src/utils/file-handlers/index.js
@@ -11,98 +11,69 @@ import { useToasts } from "../../components/Toasts"
 import moment from "moment"
 import cloneDeep from "lodash/cloneDeep"
 import fromUDTCSV from "../from-udt-csv.js"
-import usePosthog from "../../utils/use-posthog"
+import useEventCallback from "use-event-callback"
+import useFileTelemetry from "./use-file-telemetry"
 
 export default () => {
   const [file, changeFile] = useState()
   const { addToast } = useToasts()
-  const posthog = usePosthog()
 
   useFilesystem(file, changeFile)
   useLocalStorage(file, changeFile)
   useServer(file, changeFile)
 
   // Telemetry
-  const udt = file && file.content
-  useEffect(() => {
-    if (!udt || !udt.interface) return
-    posthog.capture("interface_type", { interface_type: udt.interface.type })
-    posthog.people.set({ last_used_interface_type: udt.interface.type })
-  }, [udt && udt.interface && udt.interface.type])
-  useEffect(() => {
-    if (!udt || !udt.taskData) return
-    posthog.capture("dataset_size", {
-      dataset_size: (udt.taskData || []).length
-    })
-  }, [udt && (udt.taskData || []).length])
-  useEffect(() => {
-    if (!udt || !udt.taskOutput) return
-    const numCompleted = udt.taskOutput.filter(Boolean).length
-    posthog.capture("sample_completion", {
-      dataset_size: (udt.taskData || []).length,
-      samples_completed: numCompleted,
-      percent_completed: numCompleted / (udt.taskData || []).length
-    })
-  }, [
-    udt && (udt.taskData || []).length,
-    udt && udt.taskOutput && udt.taskOutput.filter(Boolean).length
-  ])
+  useFileTelemetry(file && file.content)
 
-  const openFile = useMemo(
-    () => fsFile => {
-      const { name: fileName, path: filePath } = fsFile
-      const reader = new FileReader()
-      reader.onload = e => {
-        const content = e.target.result
-        try {
-          let oha
-          if (fileName.endsWith("csv")) {
-            oha = fromUDTCSV(content)
-          } else {
-            oha = JSON.parse(content)
-          }
-          // TODO validate OHA and prompt to open anyway if invalid
-          changeFile({
-            fileName,
-            filePath,
-            mode: filePath ? "filesystem" : "local-storage",
-            content: oha,
-            id: filePath
-          })
-        } catch (e) {
-          console.log(e.toString())
-          addToast("Couldn't read file, see console for details", "error")
-          // addError(`Could not read file "${fsFile.name}"`)
+  const openFile = useEventCallback(fsFile => {
+    const { name: fileName, path: filePath } = fsFile
+    const reader = new FileReader()
+    reader.onload = e => {
+      const content = e.target.result
+      try {
+        let oha
+        if (fileName.endsWith("csv")) {
+          oha = fromUDTCSV(content)
+        } else {
+          oha = JSON.parse(content)
         }
+        // TODO validate OHA and prompt to open anyway if invalid
+        changeFile({
+          fileName,
+          filePath,
+          mode: filePath ? "filesystem" : "local-storage",
+          content: oha,
+          id: filePath
+        })
+      } catch (e) {
+        console.log(e.toString())
+        addToast("Couldn't read file, see console for details", "error")
+        // addError(`Could not read file "${fsFile.name}"`)
       }
-      reader.readAsText(fsFile)
-    },
-    [changeFile]
-  )
+    }
+    reader.readAsText(fsFile)
+  })
 
-  const openUrl = useCallback(
-    async url => {
-      const sessionId = decodeURIComponent(url.match(/[\?&]s=([^&]+)/)[1])
-      if (!sessionId) return
-      const { state, version } = await getLatestState(sessionId)
-      if (!state) return
-      window.history.replaceState(
-        {},
-        window.document.title,
-        `/?s=${encodeURIComponent(sessionId)}`
-      )
-      changeFile({
-        url,
-        sessionId,
-        mode: "server",
-        id: sessionId,
-        content: state,
-        lastSyncedState: cloneDeep(state),
-        lastSyncedVersion: version
-      })
-    },
-    [changeFile]
-  )
+  const openUrl = useEventCallback(async url => {
+    const sessionId = decodeURIComponent(url.match(/[\?&]s=([^&]+)/)[1])
+    if (!sessionId) return
+    const { state, version } = await getLatestState(sessionId)
+    if (!state) return
+    window.history.replaceState(
+      {},
+      window.document.title,
+      `/?s=${encodeURIComponent(sessionId)}`
+    )
+    changeFile({
+      url,
+      sessionId,
+      mode: "server",
+      id: sessionId,
+      content: state,
+      lastSyncedState: cloneDeep(state),
+      lastSyncedVersion: version
+    })
+  })
 
   useEffect(() => {
     if (!file) return
@@ -116,7 +87,7 @@ export default () => {
     }
   }, [])
 
-  const makeSession = useCallback(async () => {
+  const makeSession = useEventCallback(async () => {
     const newFile = await convertToCollaborativeFile(file)
     changeFile(newFile)
     window.history.replaceState(
@@ -124,7 +95,7 @@ export default () => {
       window.document.title,
       `/?s=${encodeURIComponent(newFile.sessionId)}`
     )
-  }, [file, changeFile])
+  })
 
   return useMemo(() => ({ file, changeFile, openFile, openUrl, makeSession }), [
     file,
